insert product

+++ Domain/Models/Product.cs
@@ -18,7 +18,7 @@ namespace Domain.Models
 
         public string? ProductDescription { get; set; }
 
-        public int? ProductPrice { get; set;}
+        public decimal? ProductPrice { get; set;}
 
         public int? CategoryId { get; set;}
